Use joinedTeams as base folder instead of sites
The sites endpoint returns every single sharepoint site that a user has
access to. This list can include many more options than a user would
expect, including stuff that the user should technically not have access
to (such as if a site has its permissions misconfigured). While we can
use a $search param to narrow the results, we can only get so granular
and so even the filtered results may provide an unwieldy amount of
folders.

Because IU is interested in uploading files that are stored on
sharepoint drives that originate from Microsoft Teams, we can use the
`joinedTeams` endpoint to only return drives that a user is explicitly a
member of. This creates a much more sensible and manageable list for
users to browse through and one that should have no surprises for the
user.

This commit also includes some Rubcocop rule management changes to get
everything returning green.